+++ spreadsheet/__tests__/overlays/column_chooser/columnItem.test.js
@@ -8,8 +8,8 @@ import { DndProvider } from "react-dnd";
 import { TouchBackend } from "react-dnd-touch-backend";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import MultiBackend, { TouchTransition } from "react-dnd-multi-backend";
-import ColumnsList from "../../../../src/overlays/column_chooser/columnsList";
-import ColumnItem from "../../../../src/overlays/column_chooser/columnItem";
+import ColumnsList from "../../../src/overlays/column_chooser/columnsList";
+import ColumnItem from "../../../src/overlays/column_chooser/columnItem";
 import "@testing-library/jest-dom/extend-expect";
 
 let container;
