+++ spreadsheet/__tests__/overlays/export_data/ExportData.test.js
@@ -6,7 +6,7 @@ import ReactDOM from "react-dom";
 import ReactTestUtils, { act } from "react-dom/test-utils";
 import { render, fireEvent, cleanup } from "@testing-library/react";
 import "@testing-library/jest-dom/extend-expect";
-import ExportData from "../../../../src/overlays/export_data/ExportData";
+import ExportData from "../../../src/overlays/export_data/ExportData";
 
 let container;
